Initial commit
Commiting basic file structure with templates for the grunt processes.
However it is yet to be determined which tasks are really needed and how
certain tasks such as the imagemin one will have to be changed to apply
to wordpress. Also it remains to be seen whether the php files will just
be copied over or if there is a task to somehow compile them.

The wp install will go into wwwroot/. Do not forget to recheck the
.gitignore after adding it. Question is whether wordpress will always be
pulled in from a remote source (latest link) when a new clone is setup.
Would probably be the cleanest way to avoid merge conflicts. Since the
whole theme folder is compiled by grunt tasks from the source folder
there will be no problems with overwriting and stuff. And since
wordpress clearly separates its core from the themes this should be the
way to go.