configure js tasks
jshint, concatenation and uglify are now set up to work with
the usual wordpress folder structure.
The concat result is directly put into the build folder to
serve as reference for the styleguide and to mirror the way
things are done for the stylesheet where the style.css has
to be in the theme folder anyways.