Setup Wordpress getter task
This task allows a user to get the latest version of Wordpress. It
should only be run if the Wordpress installation is not yet part of the
repo, i. e. when starting from scratch from the boilerplate.

We might need several files outside of the themes folder, esp. uploads
that might not exist in the src folder, since they are generated by
Wordpress and stored directly to specific folders. To make maintenance
of the repo easier the whole Wordpress install will be part of it, we
will only ignore files generated by our grunt tasks.

It is therefore crucial to commit a Wordpress patch as soon as it was
installed on the server to prevent ugly merge conflicts. Also no core
files should ever be touched directly by a developer.