feat(grunt): adjust imagemin to wordpress environment
Content images are usually uploaded through the WP interface not directly through FTP
since quite some image manipulation is happening upon upload. So we redefine the
content images task of imagemin to be used solely as a upload preparer. This means
that only style images are now minified and copied to the root folder. In order for us
to still be able to prepare images for upload the second task has been split away and
can now be run explicitly whenever one is ready to upload images. These are then compressed
and put into a temporary, gitignored folder from where they can be moved to WP through
its own uploader.

Closes #4