feat(grunt): add default htaccess and tasks to copy it
The default htaccess includes the basic wordpress rewrites and can be
improved by adding some custom rules for i. e. caching and compression.
The tasks configured are just a basic copy from the src to the dest,
including it in the defined tasks as well as adding a watch task for it.