fix(rsync): improve configuration for most use cases
We add a stagingDry task that does just a dry run on the staging server with the delete option turned on.
On our default tasks the delete option is now turned of, since changes on the server by the customer or
ourselves should not necessarily be deleted. However by running the dry task for either staging or production
we can see which files would be deleted and then decide which ones we actually want to remove from the server.

Also some old references coming over from the ported contrexx boilerplate are removed
(basically some exclude and include statements) since those files do not exist in
WordPress.