feat(grunt): automatically update devUrl on setup
The devUrl in the Gruntconfig.yml file is now automatically updated when the
`grunt setup` or `grunt setupUpdate` commands are run. The task is also available
directly through `grunt shell:updateUrl`.

Make sure to be in the root directory of your project when running each of these
commands since the script takes the name of the current working directory (pwd)
and replaces the current devUrl in the Gruntconfig with this value, also adding
the http:// and .dev.

Notice: If you are using other protocols or tdl's for your local environment you
will have to adjust the script in the `grunt/shell.js` file to reflect your own
settings.