feat(grunt): add acf pro key to composer.json on setup
The key for ACF Pro can be added to the secrets.json file and will then
automatically be added to the composer.json file where it might be
committed to the repo. If you want to keep your key out of the repo
because i. e. it is public make sure not to commit after running setup. Just use
`git checkout -- composer.json` to reset the file after running the install or an
update. This way you can keep your composer file up to date for everyone working
on the repo but still keep the key private. Of course if you are using a private
repo to start with there is less to worry about.