feat(setup): allow for wget or curl downloads
We download the keys and WordPress core we need through either wget or curl.
The script now first checks if wget is available, if not whether curl is
installed and throws an error if none can be found. Also the curl script is
slightly improved using the -o option instead of streaming to stdout and
appending it to a file with >.