feat(grunticon): enable enhanceSVG
This grunticon option allows us to dynamically embed svg's into our
markup where needed so it can be styled with CSS. Needed for this is
an updated importer script in the kss template. Also make sure to add
the latest script to your WP theme. Additionally we can remove the direct
icon CSS from the KSS config file since we now have the importer script in the
head of the kss template.