feat(wordpress): enable custom plugins and themes
There are now specific folders which get copied 1:1 into the plugins and themes directory
of Wordpress resp. This enables you to add custom/premium themes and plugins that are not
available through Composer to your site/repo. Will also be an interesting option once we
do have a starter theme and develop our child themes on top of that one.