fix(scss): less css bloat from typo definitions
Until now every rule was applied even if the variables were
exactly the same. I. e. for anchor hover: if you set the
same values for text-decoration and text-decoration-hover up
to now, both were put into the CSS even though we probably only
need one for the base tag. Same for font sizes: if the mobile
font size is the same as the desktop font size, we do not need
to repeat the rule, thus it is now ignored.