feat(grunt): add cssnext postcss plugin
This enables you to use the future CSS syntax safely today.
According to a browser specification provided by you, the
given values are transormed to older CSS as long as this is
necessary to support them. Further the references to i. e. custom
properties such as variables will not be removed but added in
after the computed value. This enables you to have fool access
to the variables from i. e. JS, which means, you are able
to change variables on the fly using JS. Also since the
variables in the code are just pointers to the definition of
itself, a change of the variable at any time results in a change
of appearance or whatever else the variable controls. Usage
example: i. e. dark/light color schemes on scroll, ...
Since cssnext includes autoprefixer we can delete its configuration
and dependency safely.