feat(js): allow standalone scripts
You now have the opportunity to keep some js partials from
being concatenated with the others, but still run through
babel and uglification as well as hashres. This allows you
to selectively enqueue rarely used scripts on just the pages
that use them. Make sure to carefully read the documentation
that is found in the src/js/no-concat directory to get a good
understanding of what is going on and on how to correctly enqueue
these files.