feat(php): optimize and cleanup the header
We move the grunticon loader to an include file which injects
it into the wp_head hook. Merge this together with the font
loader which does the same.

Additionally we add some remove_action's to cleanup the header
markup by removing the generator meta tag as well as some emoji
scripts we do not need.